Modify the English banner according to the ADR-4
