+++ banners/english/banner_var.ts
@@ -3,7 +3,7 @@ import { createVueApp } from '@src/createVueApp';
 import './styles/styles_var.scss';
 
 import BannerConductor from '@src/components/BannerConductor/BannerConductor.vue';
-import Banner from './components/BannerWrapper.vue';
+import Banner from './components/BannerCtrl.vue';
 import getBannerDelay from '@src/utils/getBannerDelay';
 import { WindowResizeHandler } from '@src/utils/ResizeHandler';
 import PageOrg from '@src/page/PageOrg';
@@ -17,10 +17,6 @@ import { createFormItems } from './form_items_var';
 import { createFormActions } from '@src/createFormActions';
 import { FormControllerVar } from './FormControllerVar';
 import { useFormModel } from '@src/components/composables/useFormModel';
-import DonationForm from '@src/components/DonationForm/Forms/MainDonationForm.vue';
-import UpgradeToYearlyForm from '@src/components/DonationForm/Forms/UpgradeToYearlyForm.vue';
-import CustomAmountForm from '@src/components/DonationForm/Forms/CustomAmountForm.vue';
-import AddressTypeForm from '@src/components/DonationForm/Forms/AddressTypeForm.vue';
 
 // Change for EN banners
 import Translations from './messages';
@@ -32,7 +28,6 @@ import { CurrencyDe } from '@src/utils/DynamicContent/formatters/CurrencyDe';
 import { OrdinalDe } from '@src/utils/DynamicContent/formatters/OrdinalDe';
 import { IntegerDe } from '@src/utils/DynamicContent/formatters/IntegerDe';
 import { DeJSONFundsContentLoader } from '@src/utils/UseOfFunds/DeJSONFundsContentLoader';
-import AlreadyDonatedContent from './content/AlreadyDonatedContent.vue';
 
 const useOfFundsContent = ( new DeJSONFundsContentLoader() ).getContent();
 
@@ -56,9 +51,7 @@ const app = createVueApp( BannerConductor, {
 	},
 	bannerProps: {
 		formController: new FormControllerVar( useFormModel() ),
-		forms: [ DonationForm, UpgradeToYearlyForm, CustomAmountForm, AddressTypeForm ],
-		useOfFundsContent,
-		alreadyDonatedContent: AlreadyDonatedContent
+		useOfFundsContent
 	},
 	resizeHandler: new WindowResizeHandler(),
 	banner: Banner,
