+++ campaign_info.toml
@@ -24,8 +24,8 @@ tracking = "org-00-230202-var"
 [english]
 # name = "C22_WMDE_Test_EN_00"
 campaign_tracking = "en00-ba-230421"
-preview_link = "/wiki/Wikipedia:Hauptseite?devbanner={{banner}}&banner=B22_WMDE_local_prototype&devMode"
-#preview_url = 'https://de.wikipedia.org/wiki/Wikipedia:Hauptseite?banner={{PLACEHOLDER}}'
+preview_link = "/wiki/Main_Page?devbanner={{banner}}&banner=B22_WMDE_local_prototype&devMode"
+preview_url = 'https://en.wikipedia.org/wiki/Main_Page?banner={{banner}}'
 wrapper_template = "wikipedia_org"
 
 [english.banners.ctrl]
